feat(web): 👻 Menu | Shortcuts
- The landing header items UX were improved.
- The Panel logo is a link to the root page.
