+++ package.json
@@ -41,6 +41,7 @@
     "@tanstack/react-query-next-experimental": "5.51.11",
     "@upstash/redis": "^1.34.0",
     "@vercel/analytics": "^1.3.1",
+    "@vercel/speed-insights": "^1.0.12",
     "axios": "^1.6.0",
     "bignumber.js": "^9.1.2",
     "bs58": "^5.0.0",
