+++ web/app/[locale]/(home)/_ui/header/index.tsx
@@ -4,6 +4,7 @@ import { NextIntlClientProvider } from 'next-intl';
 
 import { Nav } from './nav';
 
+import { AvailableTreasures } from '@/ui';
 import {
   WinnersWithPortalProvider,
   GemsProvider,
@@ -14,14 +15,14 @@ export async function Header() {
   const messages = await getMessages();
 
   return (
-    <header className="fixed flex justify-between left-0 top-0 right-0 px-4 lg:px-8 py-4 z-10">
+    <header className="fixed flex justify-between items-center left-0 top-0 right-0 px-4 lg:px-8 py-4 z-10">
       <Suspense>
         <NextIntlClientProvider messages={messages}>
           <WinnersWithPortalProvider>
             <GemsProvider>
               <BountiesProvider>
                 <Nav />
-                <button className="btn btn-square btn-ghost">Log in</button>
+                <AvailableTreasures />
               </BountiesProvider>
             </GemsProvider>
           </WinnersWithPortalProvider>
