+++ web/messages/en.json
@@ -920,7 +920,7 @@
     "messages": {
       "welcome": {
         "text": "Wonderful to see you here! My name is Lucky, and I'm here to guide you through the Presale section of Luckyland.",
-        "advice": "The Presale will only be available for a limited time, so make sure to join before it's too late.",
+        "advice": "The early access is a limited offer, so make sure to join before it's too late.",
         "select": {
           "title": "What would you like to do?",
           "options": "Go to Presale, Check Tokenomics, Check Roadmap, I want to ask"
