+++ web/messages/es.json
@@ -919,10 +919,10 @@
     "messages": {
       "welcome": {
         "text": "Maravilloso verte aquí! Mi nombre es Lucky, y voy a guiarte a través de la sección de Preventa de Luckyland.",
-        "advice": "La Preventa solo estará disponible por un tiempo limitado, así que asegúrate de unirte antes de que sea demasiado tarde.",
+        "advice": "El acceso exclusivo es limitado, así que asegúrate de unirte antes de que sea demasiado tarde.",
         "select": {
           "title": "¿Qué te gustaría hacer?",
-          "options": "Ir a la Preventa, Ver la economía del projecto, Ver la planificacion, Quiero preguntar"
+          "options": "Ir a la Preventa, Ver la economía del proyecto, Ver la planificacion, Quiero preguntar"
         }
       },
       "assistant": { "text": "" }
